+++ templates/NextJs/NextJs15-basic/src/components/index.ts
@@ -1,5 +1,0 @@
-export * from './AppProvider';
-export * from './BasicLayout';
-export * from './Container';
-export * from './Footer';
-export * from './Header';